pyproject.toml
[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "DesignSpark.Cloud"
version = "22.8.0"
description = "DesignSpark Cloud support library"
authors = [{ name = "RS Components", email = "maint@abopen.com" }]
license = { file = "LICENSE" }
classifiers = [
    'Development Status :: 3 - Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Operating System :: POSIX :: Linux',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.7',
    'Topic :: Education',
    'Topic :: System :: Hardware',
    'Topic :: Software Development :: Libraries'
]
keywords = ["designspark", "cloud"]
dependencies = [
    "requests",
    "python-snappy",
    "protobuf<3.20.1",
    "prometheus-api-client"
]
requires-python = ">=3.7"

[project.urls]
Homepage = "https://github.com/DesignSparkRS/DesignSpark.Cloud"